ovn-tester/ovn_workload: Extract generic methods from Cluster class.
